% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{bin2d}
\alias{bin2d}
\title{Binning 2d points}
\usage{
bin2d(
  x,
  y,
  weight = NULL,
  nbins = c(100, 100),
  binlim,
  FUN = sum,
  num.cutoff = NULL
)
}
\arguments{
\item{x, y}{Coordinates.}

\item{weight}{Assign weight to each point. In the same order as coordinates.}

\item{nbins}{Number of bins on each axis in the order of c(nbin.x, nbin.y).}

\item{binlim}{x, y range for bin calculation, e.g. c(min(x),max(x),min(y),max(y)).}

\item{FUN}{Function for weight.}

\item{num.cutoff}{Remove bins with less than this number of points.}
}
\description{
Binning 2d points for count in each bin or calculation with weight assigned to each point.
}
\examples{
\dontrun{
x <- cellCoords[,1]
y <- cellCoords[,2]
binlim <- c(min(x),max(x),min(y),max(y))
bin2d(x=x, y=y, nbins=c(100,100), binlim=binlim)
}
}
