% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{calcDistance.euclidean}
\alias{calcDistance.euclidean}
\title{Calculate Euclidean distance}
\usage{
calcDistance.euclidean(coord1 = NULL, coord2 = NULL)
}
\arguments{
\item{coord1}{A numeric vector of coordinates for the first point.}

\item{coord2}{A numeric vector of coordinates for the other point.}
}
\description{
Extract cell coordinates by name of reduction, or fov name of the slide image.
}
\examples{
calcDistance.euclidean(coord1 = c(1,2,3), coord2 = c(4,5,6))
}
