% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{calcDistance.toLine}
\alias{calcDistance.toLine}
\title{Calculate distance of points to a line.}
\usage{
calcDistance.toLine(
  x,
  y,
  coords.df = NULL,
  p1 = NULL,
  p2 = NULL,
  k = NULL,
  b = NULL
)
}
\arguments{
\item{x}{x coordinates of the points.}

\item{y}{y coordinates of the points.}

\item{coords.df}{A data frame with x,y columns for coordinates of the two points defining the line.}

\item{p1}{A vector of two elements, the x and y coordinates of the first point defining the line.}

\item{p2}{A vector of two elements, the x and y coordinates of the second point defining the line.}

\item{k}{The slope of the line.}

\item{b}{The intercept of the line.}
}
\value{
A vector of distances.
}
\description{
Calculate the distance of points to a line defined by two points or a slope and an intercept.
}
\examples{
x <- 1:1
y <- 1:5
xy1 <- c(0,0)
xy2 <- c(5,5)
calcDistance.toLine(x, y, p1, p2)
calcDistance.toLine(x, y, k=1, b=0)
}
