% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{calcValue.gaussian}
\alias{calcValue.gaussian}
\title{Gives the density value of a Gaussian distribution}
\usage{
calcValue.gaussian(d, sd, h = NULL, scale.factor = 1000)
}
\arguments{
\item{d}{Distance to the center.}

\item{sd}{Standard deviation of the Gaussian distribution.}

\item{h}{Height of the center. If NULL, the AUC is 1.}

\item{scale.factor}{Scale up the y value to avoid long decimals.}
}
\description{
Calculate the density value of point x at distance d to the center of the Gaussian distribution with parameters sd and h
Multiple the y outside of this function if weighing the value by gene expression value of the cell.
}
\examples{
x <- seq(-5,5,0.05)
plot(x, calcValue.gaussian(d=x, sd=1))
}
