% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{createField}
\alias{createField}
\title{Create a field object}
\usage{
createField(
  coords,
  method.distr = "gaussian",
  method.d = "euclidean",
  weight = NULL,
  d.cutoff,
  ...
)
}
\arguments{
\item{coords}{A data.frame with each column represent one dimension. If there's a "value" column, move that into the weight section.}

\item{method.distr}{"gaussian" by default.}

\item{method.d}{"euclidean" by default.}

\item{weight}{A vector of the same length as coords, scaling the distribution.}

\item{d.cutoff}{Value for point further than this distance will have value 0.}

\item{...}{Other parameters. sd is the standard deviation of the normal distribtion. h is the height of the center point.}
}
\description{
Simply store the info, for getValueInField function to read.
}
\examples{
\dontrun{
coords.ifnb1 <- getCoords.transcript(srt, transcript="IFNB1", fov="UV109fov1")
findDistance.euclidean(0.00001,sd=1000)
field <- createField(coords.ifnb1, sd=1000, h=NULL, d.cutoff=2715, scale.factor=10000)
}
}
