% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_visualization.R
\name{FeaturePlot.cont.loess}
\alias{FeaturePlot.cont.loess}
\title{FeaturePlot.cont.loess}
\usage{
FeaturePlot.cont.loess(
  object,
  gene,
  continuous,
  group.by = NULL,
  split.by = NULL,
  assay = NULL,
  cols = NULL,
  smooth.span = NULL,
  smooth.se = F,
  trans = "log1p"
)
}
\arguments{
\item{object}{Seurat object.}

\item{gene}{Gene name.}

\item{continuous}{Column name in the metadata containing the continuous values for x-axis.}

\item{group.by}{Column name in the metadata to group color the points.}

\item{split.by}{Column name in the metadata to facet the plot.}

\item{assay}{Assay to use.}

\item{cols}{A named vector with group categories as names and colors as values.}

\item{smooth.span}{Only used when there are fewer than 1,000 observations. Fraction of points used to fit each local regression. Larger numbers make a smoother curve.}

\item{smooth.se}{Display standard error of the regression. By default FALSE.}

\item{trans}{Transformation of the x-axis. By default "log1p".}
}
\description{
Plot the gene expression along a continuous variable in meta.data. Smoothed by loess.
To check the gene expression pattern according to the colocalization values.
}
\examples{
\dontrun{
srt.inf <- srt[,srt$subCellType.1 \%in\% c("MC_Inf","MC_Inf2","MC")]
FeaturePlot.cont.loess(srt.inf, gene = "LYVE1", continuous = "values_CD4_subCellType2",
                       group.by = "Sample", split.by = "Disease", smooth.span = 0.9, trans="log1p")
}
}
