% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_visualization.R
\name{FeaturePlot.cont.rollmean}
\alias{FeaturePlot.cont.rollmean}
\title{FeaturePlot.cont.rollmean}
\usage{
FeaturePlot.cont.rollmean(
  object,
  gene,
  continuous,
  group.by = NULL,
  split.by = NULL,
  assay = NULL,
  cols = NULL,
  window.prop = 0.05,
  window.n = NULL,
  trans = "log1p"
)
}
\arguments{
\item{object}{Seurat object.}

\item{gene}{Gene name.}

\item{continuous}{Column name in the metadata containing the continuous values for x-axis.}

\item{group.by}{Column name in the metadata to group color the points.}

\item{split.by}{Column name in the metadata to facet the plot.}

\item{assay}{Assay to use.}

\item{cols}{A named vector with group categories as names and colors as values.}

\item{window.prop}{Proportion of data used in the sliding window for smoothing.}

\item{window.n}{Number of points in the sliding window for smoothing.}

\item{trans}{Transformation of the x-axis. By default "log1p".}
}
\description{
Plot the gene expression along a continuous variable in meta.data. Smoothed by rolling average sliding across the x-axis.
To check the gene expression pattern according to the colocalization values.
}
\examples{
\dontrun{
srt.inf <- srt[,srt$subCellType.1 \%in\% c("MC_Inf","MC_Inf2","MC")]
FeaturePlot.cont.rollmean(
  srt.inf,
  gene = "MMP9",
  continuous = "values_CD4_subCellType2",
  group.by = "Sample",
  split.by = "Disease",
  window.prop = 0.05,
  trans = "log1p"
)
}
}
