% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{findDistance.euclidean}
\alias{findDistance.euclidean}
\title{Find the distance d that the Gaussian distribution has a certain value.}
\usage{
findDistance.euclidean(value, sd)
}
\arguments{
\item{value}{The value of the Gaussian distribution at distance d.}

\item{sd}{Standard deviation of the Gaussian distribution.}
}
\description{
Find the distance d that the Gaussian distribution has a certain value.
For any points further than d, use 0 as approximate to simplify the calculation of value in field.
}
\examples{
findDistance.euclidean(value=0.01, sd=1)
}
