% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{getClickCoordinates}
\alias{getClickCoordinates}
\title{Get coordinates of clicks.}
\usage{
getClickCoordinates(plot.object)
}
\arguments{
\item{plot.object}{An object storing the plot.}
}
\value{
A data frame with x,y coordinates.
}
\description{
Display the plot and get the coordinates of clicks.
}
\examples{
\dontrun{
g <- ImageDimPlot.ssc(srt, fov = fov, group.by = NULL)
coords <- getClickCoordinates(g)
}
}
