% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{getCoords.feature}
\alias{getCoords.feature}
\title{Get the coordinates and values of features in a Seurat object}
\usage{
getCoords.feature(object, feature, image.name, assay = NULL, slot = "data")
}
\arguments{
\item{object}{Seurat object. Need to subset srt before applying this function.}

\item{feature}{The name of the gene.}

\item{image.name}{The name of the image fov in the Seurat object.}

\item{assay}{The name of the assay.}

\item{slot}{The slot of the assay.}
}
\description{
Get the coordinates of cells in an image fov together with the expression value of a gene.
}
\examples{
coords <- getCoords.feature(srt, feature="IFNB1", image.name="UV109fov1", assay="seqFISH", slot="data")
}
