% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{getSize}
\alias{getSize}
\title{Report the size of the image}
\usage{
getSize(object, fov)
}
\arguments{
\item{object}{Seurat object.}

\item{fov}{The name of the image fov in the Seurat object.}
}
\description{
Output in the order of: x.min, x.max, y.min, y.max.
}
\examples{
\dontrun{
binlim <- getSize(srt, fov = "roi1")
}
}
