% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{getValueInField}
\alias{getValueInField}
\title{Calculate the value of the field at each query point}
\usage{
getValueInField(field, coords)
}
\arguments{
\item{field}{A field object created by createField.}

\item{coords}{A data.frame with each column as a dimension.}
}
\description{
Calculate the value of the field at each query point.
}
\examples{
\dontrun{
coords.moDC <- getCoords.cell(srt, ind = srt$subcelltype == "monocyte", fov = "roi1")
values <- getValueInField(field, coords.moDC)
plot(sort(values))
plot(density(values))
}
}
