% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_analysis.R
\name{image_calcDistanceTo}
\alias{image_calcDistanceTo}
\title{Calculate the distance of cells to a shape on an image.}
\usage{
image_calcDistanceTo(object, fov, shape = "line", ...)
}
\arguments{
\item{object}{A Seurat object.}

\item{fov}{The name of the image fov in the Seurat object.}

\item{shape}{The shape of the line. Only "line" is supported for now.}

\item{...}{Other parameters for calcDistance functions.}
}
\value{
A vector of distances.
}
\description{
Calculate the distance of cells on an image to a defined shape such as a line.
}
\examples{
g <- ImageDimPlot.ssc(srt, fov = fov, group.by = NULL)
coords <- getClickCoordinates(g)
d <- image_calcDistanceTo(srt, fov = fov, coords.df = coords)
}
