% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_visualization.R
\name{ImageDimPlot.path}
\alias{ImageDimPlot.path}
\title{Plot the path on top of a image plot.}
\usage{
ImageDimPlot.path(
  ggplot.object,
  path.coords,
  ends.close = F,
  path.col = "red",
  path.size = 0.5,
  path.alpha = 1
)
}
\arguments{
\item{ggplot.object}{A ggplot object.}

\item{path.coords}{A data frame with x and y coordinates of the path.}

\item{ends.close}{If TRUE, close the path by connecting the last point to the first point.}

\item{path.col}{Color of the path.}

\item{path.size}{Size of the path.}

\item{path.alpha}{Transparency of the path.}
}
\value{
A ggplot object with the path added.
}
\description{
Plot the path on top of a image plot.
}
\examples{
g <- ImageDimPlot.ssc(srt, fov = fov, group.by = NULL)
coords <- getClickCoordinates(g)
ImageDimPlot.path(g, path.coords = coords)
}
