% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_visualization.R
\name{ImageDimPlot.sizeGuide}
\alias{ImageDimPlot.sizeGuide}
\title{ImageDimPlot.sizeGuide}
\usage{
ImageDimPlot.sizeGuide(
  object,
  fov,
  group.by = NULL,
  size = 0.2,
  cols = NULL,
  alpha = 1,
  radii = c(100, 200, 300, 400, 500, 700, 1000),
  gap = NULL,
  position = "bottomleft",
  direction = "v",
  shape.col = "lightblue",
  shape.alpha = 0.8,
  label = T,
  label.col = "red",
  label.size = 3,
  label.gap = NULL,
  dark.background = T,
  flip = F
)
}
\arguments{
\item{object}{Seurat object.}

\item{fov}{Name of the image.}

\item{group.by}{Column name in the metadata to group color the points.}

\item{size}{Size of the cells.}

\item{cols}{A named vector with group categories as names and colors as values.}

\item{alpha}{Transparency of the cells.}

\item{radii}{A vector of radii to plot.}

\item{gap}{Gap between circles.}

\item{position}{Position of the circles. "bottomleft" by default.}

\item{direction}{Aligning direction of the circles. "v" by default.}

\item{shape.col}{Color of the circles.}

\item{shape.alpha}{Transparency of the circles.}

\item{label}{Add radius label to the circles.}

\item{label.col}{Color of the label.}

\item{label.size}{Size of the label.}

\item{label.gap}{Gap between the circle and the label.}

\item{dark.background}{Black background. By default TRUE.}

\item{flip}{Flip coordinates to match the Seurat output. By default FALSE.}
}
\description{
Plot circles with different radius on ImageDimPlot.
Serve as a size guide to choose sd for plotField colocalization analysis.
}
\examples{
ImageDimPlot.sizeGuide(srt, fov = "UV238fov1", group.by = "subCellType.1", gap = 1000, position = "topleft", flip=T)
}
