% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{plot.mt}
\alias{plot.mt}
\title{Plot 2d raster density plot}
\usage{
\method{plot}{mt}(x = NULL, y = NULL, z, method = "tile")
}
\arguments{
\item{x, y}{x, y are vectors for coordinates of z (optional).}

\item{z}{A count matrix.}

\item{method}{"tile" by default, but can change to "raster".}
}
\description{
macOS Preview will do some interpolation of the raster to make the plot in pdf smoother/blurry. Adobe is fine.
Raster might have some white line/gap in between.
}
\examples{
plot.mt(wideMatrix)
}
