% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{plotContour}
\alias{plotContour}
\title{Plot the 3d contour plot}
\usage{
plotContour(df, flip = T, shape = "long")
}
\arguments{
\item{df}{A data.frame with columns x, y for coordinates z for values. Or a wide matrix with values.}

\item{flip}{Flip the plot to match seurat plot.}

\item{shape}{"long" by default, but can change to "wide".}
}
\description{
Plot the 3d contour plot.
}
\examples{
g <- plotContour(df.long, shape = "long")
htmlwidgets::saveWidget(rglwidget(), "field_IFNB1_nbins100_contour.html")
browseURL("field_IFNB1_nbins100_contour.html")
}
