% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{plotField}
\alias{plotField}
\title{Get df to plot the field}
\usage{
plotField(field, nbins, binlim, shape = "long")
}
\arguments{
\item{field}{A field object created by createField.}

\item{nbins}{Number of bins on each axis. If only one number, the same number is used for both x and y.}

\item{binlim}{x, y range for bin calculation, e.g. c(min(x),max(x),min(y),max(y)).}

\item{shape}{"long" by default, but can change to "wide".}
}
\description{
Calculate the value of the field over the 2d space. The result can be in long or wide format for downstream plotting.
}
\examples{
binlim.srt <- getSize(srt, image.name = "UV109fov1")
df.wide <- plotField(field, nbins = 100, binlim = binlim.srt, shape = "wide")
plot.mt(z=df.wide)
df.long <- plotField(field, nbins = 100, binlim = binlim.srt, shape = "long")
colnames(df.long) <- c("x","y","z")
plot.mt.long(df.long)
}
