% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_visualization.R
\name{plotMt.long}
\alias{plotMt.long}
\title{Plot 2d raster plot}
\usage{
plotMt.long(df, flip = F)
}
\arguments{
\item{df}{A data.frame with columns x, y, z for coordinates x,y and value z.}

\item{flip}{Flip the plot to match seurat plot.}
}
\description{
Plot 2d raster plot with data.frame in long format
}
\examples{
colnames(df.long) <- c("x","y","z")
plotMt.long(df.long)
}
