% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_plot.R
\name{plotTransition}
\alias{plotTransition}
\title{Transition animation between plots of cells}
\usage{
plotTransition(
  srt,
  initial = "umap",
  final = "images",
  group.by = NULL,
  fov = NULL,
  ncol = NULL,
  fov.size = NULL,
  point.size = 0.5,
  point.alpha = 1,
  image.coord.flip = T,
  match.scale.by = "y"
)
}
\arguments{
\item{initial}{Name of cell coordinates to start with. Embedding name in srt@reductions, or "images" for spatial slide.}

\item{final}{Name of cell coordinates to transition into.}

\item{group.by}{The name of metadata column to color by.}

\item{fov}{Slides to include. "all" or a vector of slide names.}

\item{ncol}{Number of fovs in a row.}

\item{fov.size}{c(width, height) to specify the size of fov panel grid.}

\item{point.size}{Point size for cells.}

\item{point.alpha}{Point transparency.}

\item{image.coord.flip}{To match the Seurat output, the x and y axises are by default flipped.}

\item{match.scale.by}{Scale the initial and final coordinates. NULL, x, y, or both.}

\item{object}{Seurat object.}
}
\description{
Create a transition animation between different cell coordinate plots, e.g. fov, umap, pca
}
\examples{
g <- plotTransition(srt, initial="umap", final="images", group.by="subcelltype", fov=NULL, point.size = 0.3, match.scale.by = "y")
animate(g, duration = 7, fps = 10, start_pause = 10, end_pause = 10,
        width = 16, height = 8, units = "in", res = 150,
        renderer = gifski_renderer())
anim_save("transition_animation_umap_to_images.gif")
}
