% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_visualization.R
\name{plotTransition.terminal}
\alias{plotTransition.terminal}
\title{Plot the initial frame and the final frame of the transition animation created by plotTransition.}
\usage{
plotTransition.terminal(
  srt,
  initial = "umap",
  final = "images",
  group.by = NULL,
  fov = NULL,
  ncol = NULL,
  fov.size = NULL,
  point.size = 0.5,
  point.alpha = 1,
  image.coord.flip = T,
  match.scale.by = "y"
)
}
\arguments{
\item{srt}{Seurat object.}

\item{initial}{Name of cell coordinates to start with. Embedding name in srt@reductions, or "images" for spatial slide.}

\item{final}{Name of cell coordinates to transition into.}

\item{group.by}{The name of metadata column to color by.}

\item{fov}{Slides to include. "all" or a vector of slide names.}

\item{ncol}{Number of fovs in a row.}

\item{fov.size}{c(width, height) to specify the size of fov panel grid.}

\item{point.size}{Point size for cells.}

\item{point.alpha}{Point transparency.}

\item{image.coord.flip}{To match the Seurat output, the x and y axises are by default flipped.}

\item{match.scale.by}{Scale the initial and final coordinates. NULL, x, y, or both.}
}
\description{
Use the same parameters as in the plotTransition that generated the animation.
}
\examples{
\dontrun{
g.list <- plotTransition.terminal(
  srt,
  initial = "umap",
  final = "images",
  group.by = "subcelltype",
  fov = NULL,
  point.size = 0.3,
  match.scale.by = "y"
)
g.list[[1]]
ggsave(
  "transition_animation_umap_to_images_start.png",
  width = 16,
  height = 8
)
g.list[[2]]
ggsave(
  "transition_animation_umap_to_images_end.png",
  width = 16,
  height = 8
)
}
}
