% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_visualization.R
\name{rollMean}
\alias{rollMean}
\title{Calculate the rolling mean vector}
\usage{
rollMean(x, y, window.prop = 0.05, window.n = NULL)
}
\arguments{
\item{x, y}{A vector of x and y values.}

\item{window.prop}{Proportion of data used in the sliding window for smoothing.}

\item{window.n}{Number of points in the sliding window for smoothing.}
}
\description{
Calculate the rolling mean of y along x, with moving window for smoothing.
Return the rolling mean vector in the same order as input x.
}
\examples{
x <- 1:1000
y <- x^0.5 + rnorm(1000, sd = 100)
rollMean(x,y,window.prop=0.1)
}
