% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_visualization.R
\name{rollMeanPlot}
\alias{rollMeanPlot}
\title{plot rolling mean}
\usage{
rollMeanPlot(
  x,
  y,
  group.vec = NULL,
  split.vec = NULL,
  window.prop = 0.05,
  window.n = NULL,
  trans = NULL
)
}
\arguments{
\item{x, y}{A vector of x and y values.}

\item{group.vec}{Optional. A vector of group values for coloring.}

\item{split.vec}{Optional. A vector of split values for facetting.}

\item{window.prop}{Proportion of data used in the sliding window for smoothing.}

\item{window.n}{Number of points in the sliding window for smoothing. window.prop will be ignored if window.n is provided.}

\item{trans}{Transformation of the x-axis.}
}
\description{
Visualize the change of y along a continuous variable x, with moving window for smoothing.
The underlying function for FeaturePlot.cont.rollmean.
}
\examples{
\dontrun{
rollMeanPlot(
  x = srt.inf$values_CD4_subCellType2,
  y = srt.inf@assays$seqFISH@layers$data[rownames(srt) == "MMP9", ],
  group.vec = srt.inf$Sample,
  trans = "log1p"
)
}
}
