% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{getCoords.cell}
\alias{getCoords.cell}
\title{Get the coordinates of cells in a Seurat object}
\usage{
getCoords.cell(object, fov, meta.cols = NULL)
}
\arguments{
\item{object}{Seurat object. Need to subset srt before applying this function.}

\item{fov}{The name of the image fov in the Seurat object.}
}
\description{
Extract the coordinates of cells of an fov.
Subset the Seurat object for cell types etc before applying this function.
}
\examples{
coords <- getCoords.cell(srt[,srt$CellType == "Keratinocyte"], fov = "UV109fov1")
}
