% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization.R
\name{getSize}
\alias{getSize}
\title{Report the size of the image}
\usage{
getSize(object, image.name)
}
\arguments{
\item{object}{Seurat object.}
}
\description{
Output in the order of: x.min, x.max, y.min, y.max.
}
\examples{
binlim <- getSize(srt, image.name = "roi1")
}
