% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_plot.R
\name{ImageDimPlot.ssc}
\alias{ImageDimPlot.ssc}
\title{Visualize categorical metadata and transcripts in spatial context (centroids).}
\usage{
ImageDimPlot.ssc(
  object,
  fov,
  group.by = NULL,
  split.by = NULL,
  size = 0.1,
  cols = NULL,
  alpha = 1,
  highlight.by = NULL,
  highlight.groups = NULL,
  highlight.size = 0.2,
  highlight.cols = NULL,
  highlight.alpha = 1,
  molecules = NULL,
  molecules.size = 0.1,
  molecules.cols = NULL,
  molecules.alpha = 1,
  dark.background = T,
  crop = NULL,
  flip_xy = F
)
}
\arguments{
\item{object}{Seurat object.}

\item{fov}{The fov to plot.}

\item{group.by}{The name of metadata column to color by.}

\item{split.by}{The name of metadata column to split the plot by.}

\item{size}{Point size for cells when plotting centroids.}

\item{cols}{A named vector with group categories as names and colors as values.}

\item{alpha}{Point transparency.}

\item{highlight.by}{The name of metadata column containing the highlight category labels. Same as group.by by default.}

\item{highlight.groups}{A vector specifying categories to highlight.}

\item{highlight.size}{Point size for highlighted cells.}

\item{highlight.cols}{A named vector with highlight category labels as names and colors as values.}

\item{highlight.alpha}{Transparency of highlighed cell points.}

\item{molecules}{Molecules to plot the location.}

\item{molecules.size}{Point size of molecules.}

\item{molecules.cols}{A named vector with molecules as names and colors as values.}

\item{molecules.alpha}{Transparency of molecules to plot.}

\item{crop}{If TRUE, crop off the region with no cells or molecules to plot. Or a vector with four numbers specifying c(min.x, max.x, min.y, max.y)}

\item{flip_xy}{To match the Seurat output, the x and y axises are by default flipped.}

\item{dark.backgrount}{Black backgrount. By default TRUE.}
}
\description{
Visualize categorical metadata and transcripts in spatial context (centroids).
}
\examples{
ImageDimPlot.ssc(srt, fov = "UV109fov1", group.by = "celltype_res0.1")
ImageDimPlot.ssc(srt, fov = "UV109fov1", group.by = "celltype_res0.1", split.by = "Disease")
ImageDimPlot.ssc(srt, fov = "UV109fov1", group.by = "celltype_res0.1", highlight.groups = c("Lymphocyte","KC_glandular_eccrine"))
srt$lowCount <- ifelse(srt$nCount_seqFISH < 100, "low", "high")
ImageDimPlot.ssc(srt, fov = "UV109fov1", group.by = "celltype_res0.1", highlight.by = "lowCount", highlight.groups = "low", highlight.size = 0.5)
ImageDimPlot.ssc(srt, fov = "UV109fov1", group.by = "celltype_res0.1", molecules = c("IFNB1","IL17A"), molecules.size = 0.7)
ImageDimPlot.ssc(srt, fov = "UV109fov1", group.by = "celltype_res0.1", crop = T)
}
