% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colocalization_visualization.R
\name{ImageDimPlot.values}
\alias{ImageDimPlot.values}
\title{ImageDimPlot color by provided values.}
\usage{
ImageDimPlot.values(
  coords,
  values,
  alpha = 1,
  size = 0.8,
  flip = F,
  color = c("grey", "purple", "red", "orange", "yellow"),
  srt = NULL,
  image.name = NULL,
  alpha.srt = 0.1,
  size.srt = 0.1,
  color.srt = "lightblue",
  dark.background = T
)
}
\arguments{
\item{coords}{A data.frame with columns x, y.}

\item{values}{A vector of values for coloring.}

\item{alpha}{Transparency of the points.}

\item{size}{Size of the points.}

\item{flip}{Flip the plot to match seurat plot.}

\item{color}{Color of the points.}

\item{object}{Seurat object.}
}
\description{
Plot points on a 2d space color by values. If Seurat object is provided, plot the location of the cells beneath.
}
\examples{
ind.moDC <- srt$subcelltype == "monocyte"
coords.moDC <- getCoords.cell(srt, ind = ind.moDC, image.name = "roi1")
values <- getValueInField(field, coords.moDC)
ImageDimPlot.values(coords=coords.moDC, values, image.name="roi1", srt=srt)

}
